Extract Windows fn from self-update.rs
Continues the strategy as described,
moving all Windows functionality to windows.rs
